requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain_ollama
langchain_openai
pandas 
python-multipart
sqlalchemy 
aiosqlite
langchain-community